+++ src/components/item/Header/Header.js
@@ -1,10 +1,8 @@
 import React, {Component} from 'react';
-import {
-  StyleSheet,
-  SafeAreaView,
-} from 'react-native';
+import {StyleSheet, SafeAreaView} from 'react-native';
 import {Actions} from 'react-native-router-flux';
 import Animated, {Easing, interpolate} from 'react-native-reanimated';
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 
 import appConfig from 'app-config';
 
@@ -12,6 +10,9 @@ import Container from '../../Layout/Container';
 import OverlayIconButton from './OverlayIconButton';
 import RightButtonNavBar from '../../RightButtonNavBar';
 import {RIGHT_BUTTON_TYPE} from '../../RightButtonNavBar/constants';
+import {CONFIG_KEY, isConfigActive} from 'src/helper/configKeyHandler';
+
+const AnimatedIcon = Animated.createAnimatedComponent(MaterialCommunityIcons);
 
 const styles = StyleSheet.create({
   container: {
@@ -54,6 +55,9 @@ const styles = StyleSheet.create({
 const ACTIVE_OFFSET_TOP = 100;
 const RIGHT_BUTTONS = [
   {
+    type: RIGHT_BUTTON_TYPE.WAREHOUSE,
+  },
+  {
     type: RIGHT_BUTTON_TYPE.CHAT,
     iconName: 'ios-chatbubbles',
   },
@@ -70,6 +74,7 @@ const RIGHT_BUTTONS = [
 class Header extends Component {
   static defaultProps = {
     item: {},
+    onPressWarehouse: () => {},
   };
 
   state = {
@@ -78,38 +83,29 @@ class Header extends Component {
   animatedOpacity = new Animated.Value(0);
   animatedTranslate = new Animated.Value(0);
   animatedScale = new Animated.Value(1);
+  animatingScrollDown = false;
+  animatingScrollUp = false;
+  animationDuration = 250;
 
-  get iconName() {
-    return appConfig.device.isIOS ? 'ios-arrow-back' : 'md-arrow-back';
-  }
-
-  get iconBackgroundStyle() {
-    return {
-      opacity: interpolate(this.animatedOpacity, {
-        inputRange: [0, 1],
-        outputRange: [1, 0],
-      }),
-    };
-  }
-
-  get iconStyle() {
-    return {
-      opacity: this.animatedOpacity,
-    };
-  }
-
-  get iconOverlayStyle() {
-    return {
-      transform: [
-        {
-          scale: interpolate(this.animatedScale, {
-            inputRange: [0, 1],
-            outputRange: [1, 1.1],
-          }),
-        },
-      ],
-    };
-  }
+  iconStyle = {opacity: this.animatedOpacity};
+  iconOverlayStyle = {
+    transform: [
+      {
+        scale: interpolate(this.animatedScale, {
+          inputRange: [0, 1],
+          outputRange: [1, 1.1],
+        }),
+      },
+    ],
+  };
+  iconBackgroundStyle = {
+    opacity: interpolate(this.animatedOpacity, {
+      inputRange: [0, 1],
+      outputRange: [1, 0],
+    }),
+  };
+  iconName = appConfig.device.isIOS ? 'ios-arrow-back' : 'md-arrow-back';
+  contentOverlayStyle = [this.iconBackgroundStyle, this.iconOverlayStyle];
 
   componentDidMount() {
     this.props.animatedValue.addListener(this.scrollListener);
@@ -119,9 +115,6 @@ class Header extends Component {
     this.props.animatedValue.removeListener(this.scrollListener);
   }
 
-  animatingScrollDown = false;
-  animatingScrollUp = false;
-  animationDuration = 250;
   scrollListener = ({value}) => {
     if (value > ACTIVE_OFFSET_TOP) {
       if (this.animatingScrollDown) return;
@@ -163,8 +156,41 @@ class Header extends Component {
     });
   };
 
+  handlePressWarehouse = () => {
+    this.props.onPressWarehouse();
+  };
+
+  renderWarehouseIcon = (iconStyle) => {
+    return (
+      <Animated.View>
+        <AnimatedIcon name="warehouse" style={[iconStyle, this.iconStyle]} />
+      </Animated.View>
+    );
+  };
+
+  renderOverlayWarehouseIcon = (
+    iconStyle,
+    iconOverlayStyle,
+    contentOverlayStyle,
+  ) => {
+    return (
+      <Animated.View style={[contentOverlayStyle, this.contentOverlayStyle]}>
+        <AnimatedIcon
+          name="warehouse"
+          style={[iconStyle, iconOverlayStyle, this.iconOverlayStyle]}
+        />
+      </Animated.View>
+    );
+  };
+
   renderRightButtons() {
     return RIGHT_BUTTONS.map((btn, index) => {
+      if (
+        btn.type === RIGHT_BUTTON_TYPE.WAREHOUSE &&
+        !isConfigActive(CONFIG_KEY.SELECT_STORE_KEY)
+      ) {
+        return null;
+      }
       const isLast = index === RIGHT_BUTTONS.length - 1;
 
       const narrowRightGapStyle = {
@@ -181,16 +207,28 @@ class Header extends Component {
             type={btn.type}
             shareTitle={this.props.item.name}
             shareURL={this.props.item.url}
+            onPress={
+              btn.type === RIGHT_BUTTON_TYPE.WAREHOUSE
+                ? this.handlePressWarehouse
+                : null
+            }
             icon={
               <OverlayIconButton
                 iconName={btn.iconName}
                 containerStyle={styles.containerIconStyle}
                 backgroundStyle={this.iconBackgroundStyle}
-                contentOverlayStyle={[
-                  this.iconBackgroundStyle,
-                  this.iconOverlayStyle,
-                ]}
+                contentOverlayStyle={this.contentOverlayStyle}
                 iconStyle={this.iconStyle}
+                renderMainIcon={
+                  btn.type === RIGHT_BUTTON_TYPE.WAREHOUSE
+                    ? this.renderWarehouseIcon
+                    : null
+                }
+                renderOverlayIcon={
+                  btn.type === RIGHT_BUTTON_TYPE.WAREHOUSE
+                    ? this.renderOverlayWarehouseIcon
+                    : null
+                }
                 disabled
               />
             }
@@ -234,10 +272,7 @@ class Header extends Component {
             <OverlayIconButton
               iconName={this.iconName}
               backgroundStyle={this.iconBackgroundStyle}
-              contentOverlayStyle={[
-                this.iconBackgroundStyle,
-                this.iconOverlayStyle,
-              ]}
+              contentOverlayStyle={this.contentOverlayStyle}
               iconStyle={this.iconStyle}
               wrapperStyle={narrowLeftGapStyle}
               onPress={Actions.pop}